requirements.txt
fastapi
uvicorn
httpx
msal
python-dotenv
pydantic
psycopg2-binary
